+++ src/main/java/tn/esprit/spring/entities/Joueur.java
@@ -33,7 +33,7 @@ public class Joueur {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	Long idJoueur;
+	Integer idJoueur;
 	
 	String nom;
